made seeds weighted equally
I realized that stacking all the vertices from a network was causing rois
with more vertices to be more highly wighted than other rois. So now we just mean
across the sampling roi matrix.  This is very useful because it gives me the idea of intentional weigthing to use other templates..